refactor: merge `config_file_initializer.py` into `config.py`
This change centralizes the configuration logic by moving the `ConfigFileInitializer` class directly into `config.py`.
- Moved `ConfigFileInitializer` from `src/config_file_initializer.py` to `src/config.py`.
- Updated `src/display.py` to import `ConfigFileInitializer` from `config`.
- Consolidated test logic by moving tests to `tests/test_config.py` and deleting `tests/test_config_file_initializer.py`.
- Fixed circular imports through `typing.TYPE_CHECKING` for type annotations.
- Addressed code review nitpicks by removing an unused local import.
