+++ src/config.py
@@ -2,8 +2,12 @@ import configparser
 import logging
 from dataclasses import dataclass
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from issue import Status
+
+if TYPE_CHECKING:
+    from jira_client import JiraClient
 
 
 @dataclass
@@ -77,3 +81,24 @@ class Config:
         level_str = config.get('Logging', 'level', fallback='INFO').upper()
         level = getattr(logging, level_str, logging.INFO)
         return LoggingConfig(level=level)
+
+
+class ConfigFileInitializer:
+    def __init__(self, config_file_path: str):
+        self.config_file_path = config_file_path
+        self.config = configparser.ConfigParser(comment_prefixes=('#', ';'), allow_no_value=True)
+        self.config.read(self.config_file_path)
+
+    def initialize_status_mapping(self, jira_client: 'JiraClient', project_name: str):
+        statuses = jira_client.fetch_jira_statuses()
+
+        if not self.config.has_section('StatusMapping'):
+            self.config.add_section('StatusMapping')
+
+        for status in statuses:
+            if not self.config.has_option('StatusMapping', status.id):
+                self.config.set('StatusMapping', f'# {status.name}')
+                self.config.set('StatusMapping', status.id, Status.TO_DO)
+
+        with open(self.config_file_path, 'w') as configfile:
+            self.config.write(configfile)
